+++ extensions/ql-vscode/src/remote-queries/gh-api/gh-api-client.ts
@@ -94,3 +94,25 @@ export async function getRepositoryFromNwo(
   const response = await octokit.rest.repos.get({ owner, repo });
   return response.data as Repository;
 }
+
+
+/**
+ * Creates a gist with the given description and files.
+ * Returns the URL of the created gist.
+ */
+export async function createGist(
+  credentials: Credentials,
+  description: string,
+  files: { [key: string]: { content: string } }
+): Promise<string | undefined> {
+  const octokit = await credentials.getOctokit();
+  const response = await octokit.request('POST /gists', {
+    description,
+    files,
+    public: false,
+  });
+  if (response.status >= 300) {
+    throw new Error(`Error exporting variant analysis results: ${response.status} ${response?.data || ''}`);
+  }
+  return response.data.html_url;
+}
